dev branch initial commit

+++ server/routes/animal/deleteAnimal.js
@@ -1,12 +1,16 @@
 const Animal = require('../../models/animal');
 const Type = require('../../models/type');
-var updateValidation = {runValidators:true, context: 'query'};
-
+const mongoose = require('mongoose');
 const deleteAnimal = async (req, res, next)=>{
   try {
-    await Type.updateOne({animal_ids:req.params.id}, {'$pull': {animal_ids: req.params.id}}, updateValidation);
+    if( mongoose.isValidObjectId(req.params.id) === false ) throw new Error('Invalid Url.');
     var delAnimal = await Animal.deleteOne({_id:req.params.id});
-    if (delAnimal.deletedCount != 1) throw new Error('Error deleting data.');
+    if ( delAnimal.deletedCount != 1 ) throw new Error('Error deleting data.');
+    var type = await Type.findOne({animal_ids: req.params.id});
+    if ( type != null ) {
+      type.animal_ids.pull(req.params.id);
+      await type.save({validateBeforeSave: false});
+    }
     res.status(200).json({'message': 'Animal deleted.'});
   } catch (error) {
     res.status(400).json([error.message])
